Add files via upload
Aim: 
Write a program to reverse the digits of a given integer in C++ by extracting and displaying each digit in reverse order.

Theory:
The program reverses a number by repeatedly extracting its last digit using the modulus operator (%) and removing it using integer division (/). This process continues until all digits are processed. The extracted digits are then combined to form the reversed number, which is printed to the console.
Example:
For the input 4664256:
Extract 6, remove last digit: 4664256 % 10 = 6, 4664256 / 10 = 466425
Extract 5, remove last digit: 466425 % 10 = 5, 466425 / 10 = 46642
Continue this process until all digits are reversed, resulting in the output 6524664.