Add files via upload
Aim: The aim of this program is to demonstrate how to use conditional statements (if and else) in C++ to determine whether an input number is positive or negative.
Theory: This program uses conditional statements to classify a number as either positive or negative. The program begins by prompting the user to enter an integer. It then employs an if statement to check whether the entered number is greater than or equal to zero. If the condition is met, the program outputs that the number is positive. If not, the else block executes, indicating that the number is negative.