Add files via upload
Aim: To determine the quadrant in which a point (represented by coordinates a and b) lies on a Cartesian plane based on the input values of these coordinates.
Theory: This program determines which quadrant a point lies in on a Cartesian coordinate system based on its coordinates. The user is prompted to enter two integer values representing the x and y coordinates. The program then uses conditional statements to classify the point into one of the four quadrants:

-1st Quadrant if both coordinates are non-negative,
-2nd Quadrant if the x-coordinate is negative and the y-coordinate is non-negative,
-3rd Quadrant if both coordinates are negative,
-4th Quadrant if the x-coordinate is non-negative and the y-coordinate is negative.
By evaluating the signs of the coordinates, the program outputs the corresponding quadrant.